% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censusDataLoader.R
\name{setCancensusApi}
\alias{setCancensusApi}
\title{Setting the cancensus API key}
\usage{
setCancensusApi(apiKey, overWrite = FALSE)
}
\arguments{
\item{apiKey}{for user.}

\item{overWrite}{if want to overwrite the existing API in the system.}
}
\value{
nothing
}
\description{
Setting the cancensus API key
}
\examples{
setCancensusApi(apiKey)
}
